Refactor CSV export to move business logic to entity class
